#642 relocate error page to pages from components
