+++ src/frontend/pages/ChangeRequestDetailPage/ReviewChangeRequestModal/review-change-request.tsx
@@ -6,7 +6,7 @@
 import { useParams } from 'react-router-dom';
 import { useAuth } from '../../../../services/auth.hooks';
 import { useReviewChangeRequest } from '../../../../services/change-requests.hooks';
-import ErrorPage from '../../../components/error-page/error-page';
+import ErrorPage from '../../../pages/ErrorPage/error-page';
 import LoadingIndicator from '../../../components/loading-indicator/loading-indicator';
 import ReviewChangeRequestsView from './review-change-request/review-change-request';
