+++ src/frontend/pages/ChangeRequestDetailPage/change-request-details.tsx
@@ -8,7 +8,7 @@ import { useSingleChangeRequest } from '../../../services/change-requests.hooks'
 import { useAuth } from '../../../services/auth.hooks';
 import ChangeRequestDetailsView from './change-request-details/change-request-details';
 import LoadingIndicator from '../../components/loading-indicator/loading-indicator';
-import ErrorPage from '../../components/error-page/error-page';
+import ErrorPage from '../../pages/ErrorPage/error-page';
 import './change-request-details.module.css';
 
 const ChangeRequestDetails: React.FC = () => {
