+++ src/frontend/pages/ChangeRequestsPage/change-requests-table.tsx
@@ -15,7 +15,7 @@ import CRTable from './change-requests-table/change-requests-table'; // Directly
 import ChangeRequestsFilter from './change-requests-filter/change-requests-filter';
 import LoadingIndicator from '../../components/loading-indicator/loading-indicator';
 import ActionButton from '../../components/action-button/action-button';
-import ErrorPage from '../../components/error-page/error-page';
+import ErrorPage from '../../pages/ErrorPage/error-page';
 import PageTitle from '../../components/page-title/page-title';
 import styles from './change-requests-table.module.css';
