+++ src/frontend/pages/ProjectDetailPage/project-page.tsx
@@ -9,7 +9,7 @@ import { useSingleProject } from '../../../services/projects.hooks';
 import ProjectViewContainer from './project-view-container/project-view-container';
 import ProjectEditContainer from './project-edit-container/project-edit-container';
 import LoadingIndicator from '../../components/loading-indicator/loading-indicator';
-import ErrorPage from '../../components/error-page/error-page';
+import ErrorPage from '../../pages/ErrorPage/error-page';
 import './project-page.module.css';
 
 interface ProjectPageProps {
