+++ src/frontend/pages/WorkPackageDetailPage/activate-work-package-modal-container/activate-work-package-modal-container.tsx
@@ -9,7 +9,7 @@ import { useAuth } from '../../../../services/auth.hooks';
 import { useCreateActivationChangeRequest } from '../../../../services/change-requests.hooks';
 import { useAllUsers } from '../../../../services/users.hooks';
 import { routes } from '../../../../shared/routes';
-import ErrorPage from '../../../components/error-page/error-page';
+import ErrorPage from '../../../pages/ErrorPage/error-page';
 import LoadingIndicator from '../../../components/loading-indicator/loading-indicator';
 import ActivateWorkPackageModal from './activate-work-package-modal/activate-work-package-modal';
