+++ src/frontend/pages/WorkPackageDetailPage/work-package-page.tsx
@@ -8,7 +8,7 @@ import { WbsNumber } from 'utils';
 import { useSingleWorkPackage } from '../../../services/work-packages.hooks';
 import { useAuth } from '../../../services/auth.hooks';
 import LoadingIndicator from '../../components/loading-indicator/loading-indicator';
-import ErrorPage from '../../components/error-page/error-page';
+import ErrorPage from '../../pages/ErrorPage/error-page';
 import WorkPackageEditContainer from './work-package-edit-container/work-package-edit-container';
 import WorkPackageViewContainer from './work-package-view-container/work-package-view-container';
 import './work-package-page.module.css';
